requirements.txt
anthropic
chromadb
sentence-transformers
rank_bm25
click
rich
python-dotenv
